emotion-driven-storyteller/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
tensorflow==2.20.0
numpy==1.26.3
python-multipart
transformers
torch
pandas
scikit-learn
llama-index-core
llama-index-readers-file
requests